# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Artículos",
        "description": "API para la gestión de artículos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "E:\\Programacion2025\\gelaApiLaravel"
        }
    ],
    "paths": {
...
        "/api/people": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener la lista de personas",
                "description": "Devuelve una lista de todas las personas.",
                "operationId": "getPeopleList",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Person"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Crear una nueva persona",
                "description": "Crea un nuevo registro de persona en la base de datos y su estructura de carpetas asociada.",
                "operationId": "storePerson",
                "requestBody": {
                    "description": "Datos de la persona",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Edgar Fabian"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Gonzalez Perez"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "id_card": {
                                        "type": "string",
                                        "example": "12345678-9"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+506 88887777"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, Anytown"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane.doe@example.com"
                                    },
                                    "linkedin_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://www.linkedin.com/in/janedoe"
                                    },
                                    "facebook_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://www.facebook.com/janedoe"
                                    },
                                    "indeed_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://www.indeed.com/profiles/janedoe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/people/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener una persona por su ID",
                "description": "Devuelve los detalles de una persona específica.",
                "operationId": "getPersonById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar una persona",
                "description": "Actualiza un registro de persona existente y renombra su carpeta si el nombre ha cambiado.",
                "operationId": "updatePerson",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la persona a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Edgar Fabian"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Gonzalez Perez"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "id_card": {
                                        "type": "string",
                                        "example": "12345678-9"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+506 88887777"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, Anytown"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane.doe@example.com"
                                    },
                                    "linkedin_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://www.linkedin.com/in/janedoe"
                                    },
                                    "facebook_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://www.facebook.com/janedoe"
                                    },
                                    "indeed_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://www.indeed.com/profiles/janedoe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Eliminar una persona",
                "description": "Elimina una persona y su estructura de carpetas asociada.",
                "operationId": "deletePerson",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona eliminada exitosamente"
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Person": {
                "title": "Persona",
                "description": "Modelo de una persona",
                "properties": {
                    "id": {
                        "description": "ID de la persona",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "description": "Primer nombre de la persona",
                        "type": "string",
                        "example": "Jane"
                    },
                    "last_name": {
                        "description": "Apellido de la persona",
                        "type": "string",
                        "example": "Doe"
                    },
                    "date_of_birth": {
                        "description": "Fecha de nacimiento",
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "id_card": {
                        "description": "Número de identificación",
                        "type": "string",
                        "example": "12345678-9"
                    },
                    "phone_number": {
                        "description": "Número de teléfono",
                        "type": "string",
                        "example": "+50688887777"
                    },
                    "address": {
                        "description": "Dirección",
                        "type": "string",
                        "example": "123 Main St, Anytown"
                    },
                    "email": {
                        "description": "Correo electrónico",
                        "type": "string",
                        "format": "email",
                        "example": "jane.doe@example.com"
                    },
                    "linkedin_url": {
                        "description": "URL de LinkedIn",
                        "type": "string",
                        "example": "https://www.linkedin.com/in/janedoe"
                    },
                    "facebook_url": {
                        "description": "URL de Facebook",
                        "type": "string",
                        "example": "https://www.facebook.com/janedoe"
                    },
                    "indeed_url": {
                        "description": "URL de Indeed",
                        "type": "string",
                        "example": "https://www.indeed.com/profiles/janedoe"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}